:recycle: Move memo component to component package
